fix: verify paypal payment before marking order as complete
