+++ .env.example
@@ -1,4 +1,7 @@
 PORT=5000
-MONGO_URI=YOUR_MONGO_URI
-PAYPAL_CLIENT_ID=YOUR_PAYPAL_CLIENT_ID
-JWT_SECRET=YOUR_JWT_SECRET
+MONGO_URI=<your_mongo_db_uri>
+JWT_SECRET=<your_secret>
+PAYPAL_CLIENT_ID=<your_paypal_client_id>
+PAYPAL_APP_SECRET=<your_paypal_secret>
+PAYPAL_API_URL=https://api-m.sandbox.paypal.com
+
