+++ backend/utils/paypal.js
@@ -1,0 +1,47 @@
+import dotenv from 'dotenv';
+dotenv.config();
+const { PAYPAL_CLIENT_ID, PAYPAL_APP_SECRET, PAYPAL_API_URL } = process.env;
+
+async function getPayPalAccessToken() {
+  const auth = Buffer.from(PAYPAL_CLIENT_ID + ':' + PAYPAL_APP_SECRET).toString(
+    'base64'
+  );
+
+  const url = `${PAYPAL_API_URL}/v1/oauth2/token`;
+
+  const headers = {
+    Accept: 'application/json',
+    'Accept-Language': 'en_US',
+    Authorization: `Basic ${auth}`,
+  };
+
+  const body = 'grant_type=client_credentials';
+  const response = await fetch(url, {
+    method: 'POST',
+    headers,
+    body,
+  });
+
+  if (!response.ok) throw new Error('Failed to get access token');
+
+  const json = await response.json();
+
+  return json.access_token;
+}
+
+export async function verifyPayPalPayment(paypalTransactionId) {
+  const accessToken = await getPayPalAccessToken();
+  const paypalResponse = await fetch(
+    `${PAYPAL_API_URL}/v2/checkout/orders/${paypalTransactionId}`,
+    {
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${accessToken}`,
+      },
+    }
+  );
+  if (!paypalResponse.ok) throw new Error('Failed to verify payment');
+
+  const { status } = await paypalResponse.json();
+  return status === 'COMPLETED';
+}
